add support for mdfind
In Mac, locate is not available by default because it requires database.
Instead of locate, you know, mdfind is available in Mac.
This commit adds support for mdfind for Mac user who doesn't use locate.

In addition, I decided to use {query} and {count} placeholder because
printf's formats must follow the order. They aren't available for mdfind
because a query of search must be before max count of candidates in
a command.